requirements.txt
Flask==3.1.3
Werkzeug==3.1.6
flask-cors==4.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
stripe==14.4.1
python-dotenv==1.0.0
supabase==2.28.0
websockets>=11,<16 